requirements.txt

Flask==3.0.1
Werkzeug==3.0.0
mysql-connector-python==8.4.0